Refactor in prep for p2p sync in shadow mode
