+++ runnable/generate_request_core.py
@@ -13,7 +13,8 @@ from vali_objects.utils.position_manager import PositionManager
 from vali_objects.utils.vali_bkp_utils import ValiBkpUtils, CustomEncoder
 from vali_objects.utils.subtensor_weight_setter import SubtensorWeightSetter
 from vali_objects.vali_dataclasses.perf_ledger import PerfLedgerManager
-from vali_objects.utils.auto_sync import AUTO_SYNC_ORDER_LAG_MS
+
+from vali_objects.utils.validator_sync_base import AUTO_SYNC_ORDER_LAG_MS
 
 # no filters,... , max filter
 PERCENT_NEW_POSITIONS_TIERS = [100, 50, 30, 0]
@@ -167,9 +168,6 @@ def generate_request_core(time_now:int) -> dict:
             f"directory for miners doesn't exist "
             f"[{ValiBkpUtils.get_miner_dir()}]. Skip run for now."
         )
-    
-    # Perf Ledger Calculations
-    perf_ledgers = PerfLedgerManager.load_perf_ledgers_from_disk()
 
     # we won't be able to query for eliminated hotkeys from challenge period
     hotkey_positions = position_manager.get_all_miner_positions_by_hotkey(
@@ -250,6 +248,8 @@ def generate_request_core(time_now:int) -> dict:
         n_positions_new += sum([len(p['orders']) for p in positions])
 
     assert n_orders_original == n_positions_new, f"n_orders_original: {n_orders_original}, n_positions_new: {n_positions_new}"
+
+    perf_ledgers = PerfLedgerManager.load_perf_ledgers_from_disk()
     final_dict = {
         'version': ValiConfig.VERSION,
         'created_timestamp_ms': time_now,
